Add Ruff for linting and formatting
Note: I'm using Ruff to do formatting here. It would be easy enough to
switch to Black if we prefer; but Ruff's formatter is _very_ close to
identical and there's something very nice about having a single,
zero-dependency tool which handles all this stuff. Plus it is very fast.

I've also configured `just check` so that it runs all the checks, rather
than stopping at the first failing one.