Initial import of asset build tooling and ui components
This attempts to extract the asset build tooling and component library
from job-server, and add it to airlock.  It isolates all this
tooling within an `assets` django app, and exposes a justfile to
interact with it.

You can build the assets with: `just assets/build`. `collectstatic` will
work as expected.

We also add the components browser view at the /ui-components. This acts
as a test to see everything works, and a helpful builtin reference for
using the slippers components.

Some of the components are a job-server specific, e.g. header and
footer, but are preserved as is for now.

We can probably improve on this modularity over time, but this should
allow us to use the UI components as is.