Remove the need to run collectstatic
Django's collectstatic is awkward, and unecessary in most contexts,
including this one.

This adds whitenoise to serve static files (which we would have added
anyway), and then enables its USE_FINDERS config to just serve the
assets straight from the configured STATIC_DIRS, and removes the need
for any reason to run collectstatic. This improves DX, but also
simplifies docker builds.

By default, django-vite expects the manifest to be put in STATIC_ROOT by
collectstatic. Instead, we tell it explicitly where the manifest is.