Simplify creation of default `.env` file
Given the fact that a `.env` file created in the middle of a `just` run
won't be loaded during that run, we're always going to need a minimum of
two `just` invocations when starting with a fresh environment. We used
to handle this by making the first invocation error and then requiring
that it be run again. However, it's simpler and more friendly to assume
that `devenv` will be run first and allow it to do the necessary set up.

If someone doesn't do this (e.g. they run `just test` in a completely
fresh environment) then their first invocation will fail with a message
about environment variables not being set. Subsequent invocations will
run without error. I think that this is acceptable.

This also means that the first step in each CI action becomes a bit more
natural and a bit less "One Weird Trick" i.e. running `just devenv`
rather than copying a mysterious file.