Use positional arguments for argument forwarding
Enabling the `positional-arguments` option causes the usual `$0`, `$1`,
`$2` etc environment variables to be set when running recipes. This
allows us to use the Bash `"$@"` idiom when forwarding arguments to
command which means that arguments with spaces in are handled correctly.

As an example, this now works correctly:

    just test -k 'not functional'

Note that this is in _addition_ to the `{{ arg_name }}` argument
expansion which Just provides. We recommend continuing to named
arguments for clarity and use positional arguments only for argument
forwarding.