Refactor PathItem to have plugable root directory.
This way, it can be used for viewing both workspaces and output
requests. A PathItem now takes a reference to container object, which
must have a root path, and a way to generate urls.

Added a top level Workspace object, as the first implementation of such
a Container, will add one for an OutputRequest later.

A workspace is now explict, rather than embedded in that path. As such,
I took the opportunity to explicitly add it as url parameter. The
PathItem.relpath is now always relative to the workspace directory, not
WORKSPACE_DIR.

In anticipation of adding a request file browser view, I renamed the
urls/views to be "workspace".