Basic UI buttons, and tweaks
This adds 4 conditional buttons that are displayed based on various
logic and contexts.  They do nothing, currently.

The add/remove button depends on whether you are looking at a workspace
or a request.  The submit/release depends on whether you are the author
or the request or an output checker.

We have now way to track authorship atm, it defaults to not being the
request author. To view it as a request author, just add `?is_author=t`
to the url.