Remove support for navigating from WorkspaceRoot object
Now the index page is a separate template, we don't need to support the
file-brower traversal from WorkspaceRoot, so remove that logic and
templating.  Now, file browser only supports looking at a single
workspace/request, not at a list of workspaces/requests.

In the process, improved some of the testing.