Install chromium dependencies in docker dev image
Installing these need root permissions, so need to be done
in the build. Installing the browser itself is done if/when the
functional tests actually run, and it needs to install WITHOUT
dependencies there because they're already in the image, and the
non-root user doesn't have permission to do it.