Add a functional tests helper for a logged in page
We don't need to test the full login flow again, but we will need
logged in users for future functional tests. This adds a helper
fixture, `logged_in_page`, to handle the logging in and setting
expected session data for a specific user.

The workspace_index_page tests don't test anything special, functional-
test wise, they just test that the fixture works to load a page as
an expected logged in user.