Add a functional tests fixtures for a logged in users
We don't need to test the full login flow again, but we will need
logged in users for future functional tests. This adds a helper
fixture, `login_as_user`, and 2 fixtures using it to mock an
output-checker and a researcher user, to create session data and
set expected cookies in the browser context.

The workspace_index_page tests don't test anything special, functional-
test wise, they just test that the fixture works to load a page as
an expected logged in user.