Configure WAL journal mode for SQLite database
This is needed to support a reasonable level of concurrent writes, which
isn't an issue now but is best to address before it becomes an issue.

This also gives us a framework for changing other SQLite configuration
options like `PRAGMA synchronous = normal` if we choose to do so.

Attaching a signal handler during the settings import is a _bit_
non-standard (normally you'd do it within an app's `ready()` method).
But this is a system-wide setting so there's not an obvious app to
attach it to and it more naturally belongs with the database config. I
can't thing of any specific problems having it here would cause us.