Explict view urls
The lack of top level separation in our url patterns was making it
difficult to add new url patterns, as the top level urls would match
first.

This adds an explicit "/view/" url components, so we can more easily
partition urls.

Whith this in place, it allows us to have nicer urls for the add file to
request and release files urls.

Also, it chaanges the add file url to be under workspaces, as you are
adding it from a workspace to a not known at the time request id.

This was fixed because when trying to addadditional urls for submitting
and rejectin requests, it showed up the problem.  Those urls will
benefit from this change.