Remove workspace from request url paths
Initially, we had workspace in the url to be able to more easily look up
workspace dir, and then request on disk.

However, when implementing database tracking for requests, the primary
source of truth is the db, not what's on disk.

So it seemed cleaner to just refer to requests by id alone, rather than
workspace. We still validate the the workspace exists and the user can
view it, but we will look up the workspace from the request data.

Currently, of course, we don't have a db, so this change temporarily
performs a file search to look up the request id, and infers the
workspace from there.

This step aligns the current urls and api with changes that are WIP.