Major refactor or the main API
The current workspace API mixes concerns. It was originally designed as
a template API for navigating a directory heirarchy, but we started
adding other business logic over time (creating requests, finding all
requests, releasing files). Additionaly, its not really just a workspace
API - it details with requests, and releasing to job-server.

Also, we know that we are going to have different implementations of the
API, with different storage backends (db, and job-server API).

With this is mind, I redesigned the API with the following goals:

Simple read-only plain old python objects as the main objects
 - Top level functions that act on those read-only objects
 - Group top level functions in a class
   - provides a namespace
   - some shared implementation
 - Refactor current file-only state code into a subclass of the above
   I don't expect this FileProvider to last long, but it was helpful to
   use it to prove the design.
 - separate the file browser navigation API out into its own module
   separates the design
   - we may not need this in the future anyway, makes it easy to
     remove.

I have a wip local db based implementation of this API, which is fairly
simple, as another proof of the design. But this is big change, so
I opted to try and minimise it as much as possible, but its still fairly
large.