Move selected path into container
Previously, it was in PathItem, which was awkward, especially when
making changes for file group tree view. This moves it to being an
ephemeral attribute set on the container, and thus available to all
PathIems.

I added it as an attribute, so it can be set, and to do that I switched
from a regular base class to `typing.Protocol`, which allows for
specifying attributes. The original motiviation for using a base class
(code reuse) had gone away previously