Add UrlPath type to use for all relative url paths
Previously, url paths mapped directly on to file paths, so using a full
Path for both made a lot of sense.

However, filegroups introduces a separation betweent the two. e.g. the
absolute url path "/requests/view/REQUEST_ID/default/output/file" has
a url path of `default/output/file` relative to the url root, but the
file path of `output/file` relative to the directory root.

So we use a renamed PurePosixPath as a convenient implementation of
a UrlPath.  This helps distinguish between the two types of paths via
type annotations, but also prevents accidentally doing IO operations
on the Pure path.

Also added a convenient constant ROOT_PATH, which is the empty relative
path.