Refactor url routing
Previously, we had two separately named path for workspace and requests
views.

This gives them the same name, and provides a default value for path, as
per recommend way in django docs.

This allows us to consolidate our url handling to just one reverse call,
that defaults to the root path if none is specified.