Make test factories use the api to create requests
This ensures that test requests are consistent, and have appropriate db
*and* file objects in place.

It does mean that they need to have a user associated with them, as the
api requires that, so the factories now take a user parameter. However,
by default they will construct a test user object who will have the
correct permissions to do the action requested, which is the common case
in tests.