Refactor PathItem to support groups.
The previous implementation was lazy, in that the PathItem itself
queried the file system and generated data as it was asked to.  This was
a reasonable approach when a) url paths mapped 1-1 to file paths and b)
we had a traversal based ui, we didn't need to calculate the whole tree,
only the part we were currently on.

However, we now have urls paths that have components not on disk, but in
the db (filegroups and their lists of files) that indirectly map on to
files on disk.  Trying to embed that dynamic lookup within the PathITem
was very tricky.

Also, with the tree nav UI, we need to calculate the entire tree anyway.
And generally in django, it is better to eagerly evalulate in the view
than lazily in the template.

So this change now constructs the full tree in the view, using two new
fucnctions: `get_workspace_tree` and `get_request_tree`. These
recursively walk the file-system and/or database, and construct a tree
of PathItems that are pre-filled with fixed data (e.g. children) - they
no longer look this up. As such, many things that were methods on
PathItem are now properties.

Building the tree externally has a couple of other benefits:
 - it helps separate the differences between a workspace tree and
   request tree out cleanly
 - we don't need to check for things like file existence dynamically, so
   we can remove some of those run time checks from PathItem (e.g
   exists())

There's more polish to do on the UI side of things, but it seems to work
as expected. The old traversal UI works, but there's no indictation that
a group is anything but a directory atm.