Use set_selected method rather than kwargs is_selected
Using `is_selected` to return kwargs was needed when PathItem was
frozen. However, it no longer is, so this approach, whilst still a bit
awkward, is less surprising.

The reason it is awkward is we want to set two different parameters on
the object, based on a variable that is *not* a member of the object,
and only available when we create the PathItem.