Monkeypatch pytest-playwright to allow a custom browser executable
An executable_path can be specified when launching a playwright
browser
https://playwright.dev/python/docs/api/class-browsertype#browser-type-executable-path
However, the pytest-playwright plugin doesn't provide a way to pass this
option.
(Note that the PLAYWRIGHT_BROWSERS_PATH env var seemed promising
initially, but is a way to install the browser binaries that playwright
expects to a custom path, and/or to specify that path. Playwright only
runs with the latest stable browsers by default, so we can't just point
it at a path to an old chrome version.)

This now lets us specify a custom browser executable using the
PLAYWRIGHT_BROWSER_EXECUTABLE_PATH environment variable. It also
skips playwright browser installation if this variable is set. If the
env var is not set, it'll use the default chromium browser as before.