Update to latest job-server components
This includes switching to the new modal component, which simplifies
some things. Specifically, there's now a single part to a modal
component - it now includes a button by default. It also inlines the js
wiring into the component, rather than an external script done at page
load.  This makes it work with an upcoming htmx partial rendering
change.

I change the selector in the e2e tests to match what the modal uses in
the jus to target the button.