Use htmx to render the file contents
These means we do not re-render the nav tree on every load, which
preserves the users local view state.

The entire nav tree has various hx-* attributes on it, which are
effectively inherited by all children.

Each link in the tree had a hx-get, and also an post-fetch event to
update the local tree selection state.

There's a small amount of client-side state managed here, basically
around setting the selected node, and also ensuring all details are
properly expanded. It matches the server-side generated UI.

The e2e tests pass unmodified, which is nice. They were very useful in
developing this change.

The most complex part of this was getting the modal components to work
when partially loaded with htmx. This change was made upstream in
job-server[1], and pulled into airlock in the previous commit to this.
This inlined the js needed for modals to work, meaning they would still
work when partially swap in, rather than just on full page load.

[1] https://github.com/opensafely-core/job-server/pull/4170