Handle request abspath in the api layer.
A path locating a file in a request has the form `group/path/file`, and
the file on disk is just `root/path/file`. Previously, this discrepency
was handled in file_browser_api layer, at tree construction time.

Building on the previous commit, we move this to be a responsibility of
the ReleaseRequest container. Its abspath() now expected the full path
including group, validates the group part, and returns the abspath on
disk.

Also, in the copy logic, we no longer use abspath to get the destination
path, instead using `root / path`. This means abspath can now verify the
path exists, like workspaces abspath does.

This is much cleaner, and removes some logic and attributes from
PathItem.