Optimise the partial rendering of contents.
This has two parts.

Firstly, we avoid loading the whole tree, and just load a partial tree
containing just the nodes on the selected path.  This was actually
fairly simple to do with our path walking code, by passing a list
containing just the selected_path.

Secondly, we only render the contents we need, i.e. not the tree,
headers or other parts of the UI. To do this we split out the contents
part into its own template, `file_browser/contents.html`, and just
render that.

With a large 45k node file tree, whilst the initial page load is still
slow, the subsequent partial loads are much faster.