Add factories.create_user for creating test users
In c7afc5e, we changed the the format of User.workspaces from a list to
a dict, in order to capture metadata like project.

Many permission checks worked without changes with either a list or
a dict, so many manually created test User objects still had workspaces
as list, not a dict, and tests still passed, but lead to surprising
errors.

This PR ensures that User.workspaces is a dict by adding a runtime test
to catch this error, and fixes all the places where it was incorrect.

In addition, as User.workspaces is more complex, and the project doesn't
matter for most tests, it introduces a new `factories.create_user` test
factory function, to make it easier for developers to create test Users
with the correct permissions. You can still specify a list of
workspaces, but the test factory function will ensure that they an
appropriate dict is constructed.

This makes it easy to create users that have or don't have permissions
for specific workspaces.  This is fairly important, as we've moved from
the simple "output checkers can do anything" model we started with, to
one where they still need explicit workspace permissions.