Add request permissions check into BLL
`get_release_request` now requires a user object, and checks the user
has access to the requests workspace.

This is a bit awkward in testing. We now require a correctly permissions
user to use the function to get_release_request for test assertions.
Easy enough to create one, but its a little awkward.