Implement conditional GETs for file contents
This means we will return 304s if the file contents have not changed,
which means the browser cache will be used, and we don't need to render
the content. For larger files, especially csv or text, this avoids a lot
of work.

For content that we render (text and csv atm), the etag is a bit more
involved. We want the etag value to capture both the original file and
the template file that was used. This is so that if either changes, the
cache is invalid and the new content is served.

To do this, I added a micro rendering framework that ties the templating
together with the rendering, so that we can inspect the template and
check *before* doing the potentially expensive rendering.