Convince mypy that AIRLOCK_DEV_USERS_FILE is not null
Previously the guarantee that the `AIRLOCK_DEV_USERS_FILE` setting was
not null at the point we call `read_text()` on it was implicit: we only
ever called the relevant function from a function which checked the
setting beforehand. This wasn't enough for mypy (not unreasonably) so
now we pass the value through after we check it, allowing mypy to
confirm it isn't null.