Add many more missing signatures
I had been aiming towards passing with the setting
`--disallow-untyped-calls` which didn't initially look like too much
work. But of course what happens is that the more types you add the more
the codebase falls under mypy's purview and the more problems it starts
to find.

So I'm calling time here and just commiting the changes so far.