Refactor the just added rendering framework.
In attempting to add cache busting ids to the urls, I realise that the
framework I added for rendering needed to be more general, in order to
use the same cache id logic when generating the urls.

So this refactors the rendering framework from that change into its own
standalone module, that makes it easy to use from anywhere.

`serve_file` still takes care of generating 304s when needed, but now we
can generate cache ids at url generation time too, no just when serving
the file.