Clarify is_supporting_file argument
is_supporting_file identifies whether a file is a supporting file
from a UrlPath. This is different to a ReleaseFile's relpath (it
includes the filegroup, which is part of the url for a release file,
but is not part of the relpath, which comes from its location in the
workspace).