Work around django-stubs handling of template types
Django-stubs types the return value of `get_template` as
`_EngineTemplate` – a protocol type it has made up which has only a
`render` method. We need to persuade it that we actually have a
`Template` instance with an `origin` attribute.

Also, previously we accessed `template.template.origin`. This obviously
exists, but it isn't defined on the base `Template` class. Whereas
`template.origin` is defined and gives us the same thing.