Add AuditLog model and AuditEvent BLL object
This lays the ground work for creating and reading audit log messages.
There is a new dumb BLL object AuditEvent which is created by the BLL,
and passed to the DAL to persist.

Also implemented a simple get_audit_log() API, primarily to help in
in testing, but all because we will need to show users this log at some
point.

This commit doesn't actually record any audit logs, but it sets up the
basic persistence for them.