Add UI for withdrawing files
Involved a small refactor of the supporting file logic to support more
filetype values.

Visually, in the tree, a withdrawn file is shown in italics (like
supporting file), and also with strike through.

The file itself is displayed in a similar way to supporting files, with
a notice.

Am punting on UI to allow to re-adding a withdrawn file for now.