Add views and urls for C3
This is pretty simple boilerplate.

I tried a new thing with the urls here - I explicitly match group in
the the url. I think this is much cleaner and we should probably do this
for all request urls. It will help with the confusion about whether a
UrlPath has the group in or not.