Add UI for C3
This renders the previously added forms when you navigate to group page,
as well as all the comments so far on the group.

It extends the request_view view to provide all the context needed to
display the values if you are viewing a group url.

It addes a basic happy path functional test.