Refactor file/directory detection
Previously, we used the presence of a suffix in the path to indicate
this was a file, in order to avoid having to stat() all the files, which
is slow on large workspaces.  We could do this as we enforce that all L4
files must have a file extension.

But it was a guess, and meant that an empty directory with an extension
would show up as a file, and we had to handle that.

Additionally, for incoming file browsing of git repos, we cannot enforce
that all files have an extension, so it didn't really work in that case.

However, in the git repo case, and in the request case, we only ever
have paths that are files we will never need to show empty directories.

Essentiall, all leaf nodes are files, except in the case of an empty
directory, which can only occur in the workspace tree.

So this change modifes the `get_path_tree` function to take a set of
paths that are known to be empty directofiles. It then checks membership
of this set to indicate node type. This set of paths that are known to
be empty directories is only constructed and passed to get_path_tree as
part of the workspace tree building function.

The main benefits are:
 - we do not guess file/directory type any more
 - we support empty dirs with an extension in their name
 - its faster for large workspaces.
 - it works with git repos

For workspaces, we switch to using `os.scandir`, which gives us
information on whether a path is file or directory without having to
stat() each file. It is also faster: ~1.2s on 45k node tree, vs ~1.9s
using the previous glob method, on my machine. This allows us to show
empty directories correctly in workspaces.

For requests, all paths are files, so there is no change

For workspaces, the `selected_only` logic needed to be changed to match.
This required building a file list manually if the selected file was
a directory.