Serve built docs as static files
This allows the docs to be accessed from within the same
django app, at /docs

As we're using the whitenoise option to use static file finders, we
add the built docs directory to the list of STATICFILES_DIRS in
settings, and a couple of urls to serve them using a very simple
wrapper around django's static.serve view. This lets us point
a directory level url at that directory's index.html that mkdocs
creates.