Refactor selected only logic for get_request_tree
The previous logic was more complex than it needed to be. Instead, we
generate the tree for the entire of the currently selected group. As
there will only be a handful of files in a group, this is cheap. The
additional complexity to have a slightly smaller tree was not worth it.