Refactor get_renderer/serve_file
In prepartion for serving git files, refactoring the renderer framework,
to more clearly separate obtaining a renderer, and then using it to
serve the file.

Specifically, rather than passing a ReleaseFile object into
`get_renderer`, and having it do the logic, we pass just the two
attributes we need (both of which are needed for git files).

Likewise, rather than have serve_file call get_renderer, we make the
caller pass the renderer in. This is much simpler, and makes the tests
for serve_file almost trivial.