More refactoring of rendering framework.
To render files from a git repo, we need to render from string contents
as opposed to reading from a file path.

This refactor changes the rendering framework to be based on a stream of
bytes or text, to facilitate this. This required alternate ways to
construct renderers (from_file and from_string classmethods). This
provoked a further refactor

 - the previous get_renderer function now just returns the class
 - the from_file contructor is used for workspace and release files
 - the from_string contstructor will be used for repo files
 - AirlockContainer has grown a get_renderer method, which knows how to
   look up the correct renderer class and how to instantiate it.

This is a cleaner that having get_renderer have to know about every type
of container and its files.