Add new CodeRepo container to handle git files
We import the git.py file from job-runner, with only minimal chanages
 - import changes
 - inline the project_name_from_url
 - add list_files_in_repo function

The new CodeRepo container is similar to Workspace and ReleaseRequest,
and implements AirlockContainer.

It knows how to list paths and fetch contents of files in a git commit.
We load the list of paths when creating the object, as we need to
validate the commit is valid as early as possible.

We also added a create_repo factories to create valid repos for testing