Add views for code browser
The styling is pretty basic, but it works.

It's not hooked up to any buttons yet, as we need to know the commit for
that.

To test it, you'll need to manually construction the url:

`/code/view/WORKSPACE/COMMIT`

You will also need a `metadata/manifest.json` file in the workspace, as
if generated by job-runner.  This will have a "repo" key which points to
the workspace repo.