Only run pages-deployment if docs have changed
This skips the pages-deployment workflow if there have been no changes
to any files in the docs/ folder or to mkdocs.yml. It also skips running
this workflow on the main branch, since the docs are availabe in
a deployed airlock anyway and can be checked on the test backend or in
the main docs.