Add a root node in the nav tree representing the main object.
This provides an intuiative way to navigate back to the root summary
page when browsing the tree.

The fake_parent trick is a bit odd, its basically an performance
optimisation. Without it, we'd need to recurse for every file, rather
than everthing thatn contains files. This would be a *lot* more
recursive templating, and it is not fast.

Instead, we preserve the current rendering-all--children-in-one-template
approach by making the root node have a fake parent with itself as
a child. It is a hack, but I didn't see a better way.