Send notifications on relevant events
Sends notifications on:
- set status (on moving to any status except pending)
For requests in submitted or later state):
- edit group (editing context/controls)
- add file
- withdraw file
- add comment

Currently update events are not batched, with the exception of
context and contols editing (because those are part of the same
form). The DAL now returns a list of NotificationUpdateTypes on
group_edit to indicate which fields have changed and need to be
notified.