Exclude readonly events from recent activity logs.
We log every file access for a robust audit log.

However, that adds a lot of noise when trying to understand what has
changed for a specific workspace or request.

So, we extend get_audit_log to be able to exlude readonly events, and
then use that to display a more useful recent activity.