Update login form label
The form previously asked for "GitHub username or email address". GitHub
email address is not always the email address that will work for login.
Logging in to job-server with GitHub will populate the user's job-server
account with their primary GitHub email. However, they can then change
it, and if they do, the new email address will be valid for login, not
the primary GitHub one.

We don't want to restrict users to ONLY logging in with GitHub username,
as some users may not be very familiar with GitHub and may not know
their GitHub username.