Refactor Workspace construction
This adds a manifest and current_request property, and a from_directory
contructor, which loads the manifest and any current request, and
creates the Workspace object.

This means any Workspace when created will always have a manifest, as
that is required now. This is more consistent, and uncovered a number of
issues.

It will also have any current_request for the user. This is going to be
very useful in template rendering in a coming PR

The idea is that once a workspace is instantiated, it is complete and
has everything available that it might need to then reference.