Rework directory view to use datatable
This uses a datatable to list the contents in a directory. It was spun
out of the multiselect work.

This means we can sort and order on columns, and so I also add size and
modified times for this too, to provide a use case.

There will be other columns coming soon, one for current file state
(unreleased/in-request/released for workspace files,
approved/rejected/withdrawn for request files), as well as a checkbox
column for the multiselect.