Add return urls for code view
This lets the code view called from a workspace or request file
add a return_url so the user can go back to the workspace or
specific request they came from.

If there's no return url, we show a link to the user's current
request, if there is one, or to the workspace.