Show error message and redirect bad code URLs
If a repo or commit isn't found, show an error message and redirect
user to either the return_url (which will be the current page) or
to the workspace's url. This is nicer than 404'ing from the code link
button.

For code content views, we leave it to 404, since a user would normally
be in the code browser first.