Fix load-example-data
70e19c6 included a drive-by change to allow creating workspaces from
directories that already exist.  This is used in the load-example-data
dev tooling.

Unfortunately it included the manifest writing in the directory
existing branch check, which is a logic error.  But it didn't fail at
this point, as manifest hand't yet been hard-required, it failed on
when a later change made it required.