Only include org and repo in notification if set
Previously we had a default org/repo that was sent to job-server;
now we only send these values if they're explicitly set as env
variables, and we let job-server deal with any others. This means
that the test backend can set the org/repo that should be used for
creating issues (i.e. not a real output-checking repo), and job-server
can decide which repo to use based on the organisation that the
workspace belongs to (which airlock doesn't have knowledge of)