Add a completed_reviews attribute to RequestMetadata
This is a dict that will track which users have completed a review for
the current round.  We need to know who has, so we can make sure they
cannot do two reviews.

The plan is to reset completed_reviews to empty if set the status to
RETURNED, so when it is resubmitted, any users can do the blind review
again.