Show error messages for releasing errors
From the UI, the release files view is usually an htmx request. Any
errors raised are hidden from the user. We were also returning a
TemplateResponse in debug mode for job-server errors, which was
also being swallowed in htmx mode.  This replaces the raised errors
with error messages and logging.