WIP commit of adding the REVIEWED state and UI
This was hard to split up into smaller commits, as adding and enforcing
new state transitions required new UI elements. It's mostly there, but
is lacking some coverage. It has been tested manually.

It adds two new states: PARTIALLY_REVIEWED and REVIEWED, and new view
urls and UI for users to mark their review as completed, using the above
states. It adds a new db column `completed_reviews` to track which
output checkers have completed a review for this round,
These are reset when a request is resubmitted (i.e. a new round).

Existing tests are broken in this commit - next commit is WIP fix for
them - but hopefully helpful to review.