Don't use datatables for CSVs with unequal columns
simple-datatable requires tables to have the same number of
columns in each row.

If an output CSV file contains different numbers of columns in some
rows, simple-datatables will fail to initialise the datatable. However,
a straightforward html table can be displayed.

E.g. this content can be loaded by the CSV module and displayed as an
html table, but can't be displayed by simple-datatables:
    Foo Bar
    foo,bar

Check if any rows contain different numbers of columns and pass a
context variable to the template that will prevent use of datatables.
This avoids the 5 second wait to check if the datatable can be
loaded for tables that we can verify in advance.