Update the definition of active request
We have new statuses, PARTIALLY_REVIEWED, REVIEWED and RETURNED, in which
a request is editable and should be considered active. In these
statuses, adding a file from a workspace adds it to the request if
allowed (i.e. adding a file with a request in RETURNED status adds it to
the returned request rather than creating a new one). If the user's
request is in the process of being reviewed, they can't create a new
one.