Only show the complete-review reminder where there are files to review
If there are no output files on a request, there's nothing to be
reviewed. The method for all_files_reviewed_for_user returns True,
but we don't want to remind users to complete their review when
there was nothing to review.

We also don't want to show this to authors at all.