Add Visibility enum
This enum will be used to track the current visiblity of various parts
of the system, to support both independant review and private
output-checker only comments.

BLINDED is meant to be a temporary state for votes and comments whilst
in the 2-independant review stage.  When that is completed, those
votes/comments will have their visiblity changed to PRIVATE

PRIVATE things are only viewable by output-checkers. This allows them to
leave candid feedback that the author cannot see.

PUBLIC things are viewable by all. All comments made by the author, as
well as consolidated votes/comments made by the output checkers, will be
PUBLIC