Set comment visibility via UI.
This adds a radio list to allow the user to select a comment visibility.

The visibility they are allowed to selected depends on a combination of
whether the user is the request author, an output-checker, and what
state the request is in.

When you the author or not an output-checker, the only option is PUBLIC.

When you are an output checker in the independent review phase, the only
option is BLINDED.

When you are an output checker but the request has finished the
independent review phase, you can choose PUBLIC or PRIVATE

I chose to always show the radio button list, even if there is only one
option, as it provides a help reminder as to who can see the comment you
are making.  We can always revisit this later.