Update example dev_users and check workspace keys
job-server will now send information about whether users'
workspaces are archived or not, and whether the project is
ongoing.
This is now Updated in the sample dev_users.json and addded
a check in settings.py if we're in
dev user mode to warn users early if their existing
dev_users.json hasn't been updated.