Show archived workspace/ongoing project status and order workspaces list
Workspaces are grouped by project still, but now projects are
ordered with ongoing/active projects first.

Within each project workspaces are ordered with non-archived first.

Adds a display_name() method to Workspace and Project as a template
helper for displaying the name with (ARCHIVED) or (INACTIVE) appended.
These are shown after the project/workspace name on the workspace list
view and the workspace page.

An output-checker user who doesn't have access to the workspace doesn't
have any metadata related to the workspace or project. They do stil
have access to browse workspace files, but not to create requests. In
this case we default to using "Unknown project" as the project name (as
before) and assume the project is active.